worker_pm/requirements.txt
# Database
mysql-connector-python==8.3.0

# Environment variables
python-dotenv==1.0.1

# Redis + Queue
redis==5.0.8
rq==1.16.2

# HTTP API (enqueue service)
fastapi==0.115.0
uvicorn[standard]==0.30.6
